+++ docs/changelog.rst
@@ -25,6 +25,8 @@ would have also been included in the 1.2 line.
 Changelog
 =========
 
+* :feature:`699` Allow `name` attribute on file-like objects for get/put. Thanks
+  to Peter Lyons for the pull request.
 * :bug:`711` `~fabric.sftp.get` would fail when filenames had % in their path.  
   Thanks to John Begeman
 * :bug:`702` `~fabric.operations.require` failed to test for "empty" values in